Refactor INodeFile and StdoutFile with FileHandle and SyncFile

+++ src/libos/src/fs/file_ops/lseek.rs
@@ -1,6 +1,6 @@
 use super::*;
 
-pub fn do_lseek(fd: FileDesc, offset: SeekFrom) -> Result<off_t> {
+pub fn do_lseek(fd: FileDesc, offset: SeekFrom) -> Result<usize> {
     let file_ref = current!().file(fd)?;
     file_ref.seek(offset)
 }
